Introduce coroutines framework
Adds config symbol CONFIG_COROUTINES which introduces a new internal API
for coroutines support. As explained in the Kconfig file, this is meant
to provide some kind of cooperative multi-tasking with the goal to
improve boot time by overlapping some lengthy operations.
The API as well as the implementation is very much inspired from libaco
[1]. The reference implementation has been simplified to remove all
things not needed in U-Boot, and the aco_ prefix has been replaced with
uco_ ("U-Boot coroutines").

Only i386, x86_64 and aarch64 are supported at the moment.

[1] https://github.com/hnes/libaco/

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>