Introduce coroutines framework
Add the COROUTINES Kconfig symbol which introduces a new internal API
for coroutines support. As explained in the Kconfig file, this is meant
to provide some kind of cooperative multi-tasking with the goal to
improve performance by overlapping lengthy operations.

The API as well as the implementation is very much inspired from libaco
[1] with the following differences:
- Remove all things not needed in U-Boot
- Adopte the U-Boot coding style
- Replace the aco_ prefix with co_

Support only i386, x86_64 and aarch64. Other architectures will need to
provide a _co_switch() function in assembly.

Only aarch64 has been tested.

[1] https://github.com/hnes/libaco/

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>