[RFC] efi_loader: efi_init_obj_list(): introduce coroutines
When COROUTINES is enabled, schedule efi_disks_register() and
efi_tcg2_register() as two coroutines in efi_init_obj_list() instead of
invoking them sequentially. The voluntary yield point is udelay()
which is called frequently as a result of each function polling
hardware registers or similar. The goal is to interleave the
execution of the two functions and ultimately reduce the time it
takes to run efi_init_obj_list().

Tested on KV260 ("printenv -e" command) with a microSD card inserted.
With COROUTINES disabled, efi_init_obj_list() completes in 2821 ms on
average (2825, 2822, 2817). With COROUTINES enabled, it takes 2725 ms
(2720, 2727, 2729). That is a reduction of 104 ms or 3.7%. I certainly
hoped for more because efi_tcg2_register() takes ~825 ms and
efi_disks_register() needs ~600 ms and I thought their execution would
overlap better.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>