arch: introduce symbol HAVE_INITJMP
HAVE_INIJMP will be set by architectures that support initjmp(), a
non-standard extension to setjmp()/longjmp() allowing to initialize a
jump buffer with a function pointer and a stack pointer. This will be
useful to later introduce threads. With this new function it becomes
possible to longjmp() to a particular function pointer (rather than
to a point previously reached during program execution as is the case
with setjmp()), and with a custom stack. Both things are needed to
spin off a new thread. Then the usual setjmp()/ longjmp() pair is
enough to save and restore a context, i.e., switch thread.
The initjmp() prototype is added to <asm-generic/setjmp.h> since it
is common to all architectures.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>