bootstage: add support for reporting in JSON format
Add a new Kconfig symbol: BOOTSTAGE_REPORT_JSON to report the bootstage
timing information in JSON format in addition to the human-readable
text format. JSON is easier to parse by machines and represent
graphically. The main goal is to be able to record boot statistics during
CI in order to detect performance regressions.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>