usb: fix build after resync of DWC3 with kernel v6.16-rc7
Fix build errors after the re-sync of the DWC3 driver with the kernel.
U-Boot has different needs than the kernel: buses, interrupts, internal
APIs (DMA, traces, DT...) so many adaptations are required. This commit
re-introduces many of the changes that were done locally after the
initial import of the DWC3 code from kernel 3.19-rc1 11 years ago, as
well as other fixes. This is compile-tested only.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Series-to: U-Boot mailing list <u-boot@lists.denx.de>
Series-prefix: RFC
Cover-letter:
drivers: usb: dwc3: sync code with Linux v6.16-rc1
The DWC3 USB driver was forked from the Linux kernel v3.19-rc1 eleven
years ago by commit 85d5e7075f33 ("usb: dwc3: add dwc3 folder from linux
kernel to u-boot"). Since then, not many kernel changes have been ported
back into to U-Boot. This series is an attempt at synchronizing the code
with a somewhat current kernel (namely, v6.16.rc7 which was the latest
released when this effort was started).

- The first three commits ("usb: dwc3: re-import from kernel v6.16-rc7",
  "usb: host: re-import xhci-ext-caps.h from kernel", usb: gadget:
  re-import epautoconf.c from kernel v6.16-rc7" and "usb: udc: re-import
  udc-core.c from kernel v6.16-rc7") replace files with their counterpart
  from the kernel, unmodified.
- The last commit ("usb: fix build after resync of DWC3 with kernel
  v6.16-rc7") makes the necessary changes to adapt to U-Boot and fix the
  build.

Note that this is mostly compile-tested only. The CI pipeline on
source.denx.de was used as a OK/NOK indicator. Only one config was
actually tested on real hardware: xilinx_zynqmp_kria_defconfig. With the
help of a custom build script [1] and with an additional patch [2], I
could boot the Kria KV260 board and make it detect a USB SSD plugged into
one of its USB 3.0 ports. It certainly doesn't mean all platforms using
the DWC3 driver are still OK, but at least there is some hope. If this
breaks your platform I'd like to know, and if you can send a fix it's
even better.

I realize this is a big change and I did it this way essentially to get
going and be able to post this RFC. But I am not saying this is the way
the re-sync has to be done. Feedback is welcome.

[1] https://source.denx.de/u-boot/u-boot-extras/-/raw/master/contrib/jforissier/zynqmp_kv260_build.sh
[2] https://gitlab.com/Linaro/trustedsubstrate/meta-ts/-/raw/master/meta-trustedsubstrate/recipes-bsp/u-boot/u-boot/zynqmp-kria-starter/0001-kv260-fix-usb.patch
END